Add lombok annotations

+++ exchange-library/src/main/java/ru/exlmoto/exchange/repository/BankByRepository.java
@@ -4,10 +4,10 @@ import org.springframework.data.repository.CrudRepository;
 
 import ru.exlmoto.exchange.entity.BankByEntity;
 
-public interface BankByRepository extends CrudRepository<BankByEntity, Long> {
-	BankByEntity getById(Long id);
+public interface BankByRepository extends CrudRepository<BankByEntity, Integer> {
+	BankByEntity getById(int id);
 
 	default BankByEntity getBankBy() {
-		return getById(1L);
+		return getById(1);
 	}
 }
