+++ exchange-library/src/main/java/ru/exlmoto/exchange/repository/BankKzRepository.java
@@ -4,10 +4,10 @@ import org.springframework.data.repository.CrudRepository;
 
 import ru.exlmoto.exchange.entity.BankKzEntity;
 
-public interface BankKzRepository extends CrudRepository<BankKzEntity, Long> {
-	BankKzEntity getById(Long id);
+public interface BankKzRepository extends CrudRepository<BankKzEntity, Integer> {
+	BankKzEntity getById(int id);
 
 	default BankKzEntity getBankKz() {
-		return getById(1L);
+		return getById(1);
 	}
 }
