+++ exchange-library/src/main/java/ru/exlmoto/exchange/repository/BankRuRepository.java
@@ -4,10 +4,10 @@ import org.springframework.data.repository.CrudRepository;
 
 import ru.exlmoto.exchange.entity.BankRuEntity;
 
-public interface BankRuRepository extends CrudRepository<BankRuEntity, Long> {
-	BankRuEntity getById(Long id);
+public interface BankRuRepository extends CrudRepository<BankRuEntity, Integer> {
+	BankRuEntity getById(int id);
 
 	default BankRuEntity getBankRu() {
-		return getById(1L);
+		return getById(1);
 	}
 }
