+++ exchange-library/src/main/java/ru/exlmoto/exchange/repository/BankUaRepository.java
@@ -4,10 +4,10 @@ import org.springframework.data.repository.CrudRepository;
 
 import ru.exlmoto.exchange.entity.BankUaEntity;
 
-public interface BankUaRepository extends CrudRepository<BankUaEntity, Long> {
-	BankUaEntity getById(Long id);
+public interface BankUaRepository extends CrudRepository<BankUaEntity, Integer> {
+	BankUaEntity getById(int id);
 
 	default BankUaEntity getBankUa() {
-		return getById(1L);
+		return getById(1);
 	}
 }
