+++ exchange-library/src/main/java/ru/exlmoto/exchange/repository/MetalRuRepository.java
@@ -4,10 +4,10 @@ import org.springframework.data.repository.CrudRepository;
 
 import ru.exlmoto.exchange.entity.MetalRuEntity;
 
-public interface MetalRuRepository extends CrudRepository<MetalRuEntity, Long> {
-	MetalRuEntity getById(Long id);
+public interface MetalRuRepository extends CrudRepository<MetalRuEntity, Integer> {
+	MetalRuEntity getById(int id);
 
 	default MetalRuEntity getMetalRu() {
-		return getById(1L);
+		return getById(1);
 	}
 }
